moved the method to image controller and service
